добавлено: сохранение статистики в БД и её получение по запросу

+++ stats/stats-dto/src/main/java/ru/practicum/StatsView.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class StatsView {
     private final String app;
     private final String uri;
-    private final Integer hits;
+    private final Long hits;
 }
